feature: auto-generate lens explorer site on merge (#31)
* adding a base zola site for lens site

* have validator also generate the necessary markdown for the lens explorer

* add a dry run mode to the validator

* add search files for later dev

* split explorer site generator & validator into two different commands

* generate a detail page w/ domains & urls from lens config

* add rules to detail page

* add workflow to build & deploy site

* add CNAME for lenses.spyglass.fyi

* remove test branch from deploy workflow

+++ validator/src/main.rs
@@ -1,13 +1,32 @@
 use blake2::{Blake2s256, Digest};
+use clap::{Parser, Subcommand};
 use std::ffi::OsStr;
-use std::fs;
-use std::io;
 use std::path::PathBuf;
 use std::process::ExitCode;
+use std::{fs, io};
 
 mod entity;
+mod repo;
+
 use entity::InstallableLens;
 use spyglass_lens::LensConfig;
+
+#[derive(Parser)]
+#[command(author, version, about, long_about = None)]
+pub struct ValidatorCli {
+    #[arg(short, long)]
+    pub dry_run: bool,
+    #[command(subcommand)]
+    command: Option<Commands>,
+}
+
+#[derive(Subcommand)]
+enum Commands {
+    /// Generate lens explorer site
+    GenerateExplorer,
+    /// (Default) Validates lenses & generate index file
+    Validate,
+}
 
 const INDEX_FILE: &str = "../index.ron";
 const LENS_FOLDER: &str = "../lenses";
@@ -18,20 +37,15 @@ const DL_URL_PREFIX: &str =
 fn validate_lens(lens: &InstallableLens) -> anyhow::Result<()> {
     println!("validating {} lens", lens.name);
 
-    let mut components = lens
-        .download_url
-        .split('/')
-        .collect::<Vec<&str>>();
+    let mut components = lens.download_url.split('/').collect::<Vec<&str>>();
 
-    let file = components
-        .pop()
-        .expect("Unable to get file path");
+    let file = components.pop().expect("Unable to get file path");
 
-    let parent = components.pop()
-        .expect("Unable to get parent file path");
+    let parent = components.pop().expect("Unable to get parent file path");
 
     // Attempt to parse lens
-    let contents = fs::read_to_string(format!("../lenses/{}/{}", parent, file)).expect("Unable to read lens");
+    let contents =
+        fs::read_to_string(format!("../lenses/{}/{}", parent, file)).expect("Unable to read lens");
     ron::from_str::<LensConfig>(&contents)?;
 
     Ok(())
@@ -89,9 +103,10 @@ fn check_lenses() -> io::Result<Vec<InstallableLens>> {
 
             updated_lenses.push(InstallableLens {
                 author: lens.author,
-                description: lens.description.unwrap_or_else(|| "".to_string()),
+                description: lens.description.unwrap_or_default(),
                 name: lens.name,
                 sha: hex::encode(res),
+                path: file_path.clone(),
                 download_url: format!("{}/{}/{}", DL_URL_PREFIX, parent, file_name),
                 html_url: format!("{}/{}/{}", HTML_URL_PREFIX, parent, file_name),
             });
@@ -118,21 +133,48 @@ fn validate_index_file() -> anyhow::Result<()> {
 }
 
 fn main() -> ExitCode {
-    match check_lenses() {
-        Ok(updated) => {
-            let ser = ron::ser::to_string_pretty(&updated, Default::default())
-                .expect("Unable to serialize index");
+    let cli = ValidatorCli::parse();
+    match &cli.command {
+        None | Some(Commands::Validate) => {
+            match check_lenses() {
+                Ok(updated) => {
+                    let ser = ron::ser::to_string_pretty(&updated, Default::default())
+                        .expect("Unable to serialize index");
 
-            // Write out new index file.
-            fs::write(INDEX_FILE, ser).expect("Unable to write index");
+                    // Write out new index file.
+                    if !cli.dry_run {
+                        fs::write(INDEX_FILE, ser).expect("Unable to write index");
+                    }
 
-            // Validate index file.
-            if let Err(e) = validate_index_file() {
-                println!("index validation failed: {}", e);
-                return ExitCode::FAILURE;
+                    // Validate index file.
+                    if let Err(e) = validate_index_file() {
+                        println!("index validation failed: {}", e);
+                        return ExitCode::FAILURE;
+                    }
+
+                    ExitCode::SUCCESS
+                }
+                Err(_) => ExitCode::FAILURE,
             }
-            ExitCode::SUCCESS
         }
-        Err(_) => ExitCode::FAILURE,
+        Some(Commands::GenerateExplorer) => {
+            match check_lenses() {
+                Ok(updated) => {
+                    // Generate docs
+                    println!("\ngenerating lens explorer site...");
+                    let base_path = repo::get_and_clean_doc_path(&cli);
+                    for lens in &updated {
+                        if let Err(e) = repo::generate_page(&cli, &base_path, lens) {
+                            println!("page generation failure: {}", e);
+                            return ExitCode::FAILURE;
+                        }
+                    }
+                    println!("generated {} pages", updated.len());
+
+                    ExitCode::SUCCESS
+                }
+                Err(_) => ExitCode::FAILURE,
+            }
+        }
     }
 }
